fix(bridge): move bgoffsetwrapper, fix orderhistory transition layout shift bug
